+++ src/backend/ci/core/log/api-log/build.gradle.kts
@@ -28,7 +28,6 @@
 dependencies {
     api(project(":core:common:common-api"))
     api(project(":core:common:common-web"))
-    api(project(":core:common:common-stream"))
 }
 
 plugins {
