+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/consumer/LogBatchEventListener.kt
@@ -27,7 +27,7 @@
 
 package com.tencent.devops.log.consumer
 
-import com.tencent.devops.common.log.pojo.event.LogBatchEvent
+import com.tencent.devops.log.event.LogBatchEvent
 import com.tencent.devops.log.service.BuildLogPrintService
 import com.tencent.devops.log.service.LogService
 import java.util.function.Consumer
