+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/event/LogBatchEvent.kt
@@ -25,7 +25,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.log.pojo.event
+package com.tencent.devops.log.event
 
 import com.tencent.devops.common.log.pojo.message.LogMessageWithLineNo
 import com.tencent.devops.common.stream.annotation.StreamEvent
